Seperated Controllers and added Data directory which stores team data for now
