+++ back/Data/TeamStore.cs
@@ -1,0 +1,10 @@
+using back.Models;
+using System.Collections.Generic;
+
+namespace back.Data
+{
+	public static class TeamStore
+	{
+		public static readonly List<Team> Teams = new();
+	}
+}
